✅[CHORE] : 변수 이름 변경 (#3)
폴더링 컨벤션은 좀 더 고민해보기..
